applying formatting, fixing lints
