+++ activate.zsh
@@ -1,1 +1,2 @@
 #!/bin/zsh
+export RUSTDOCFLAGS="--html-in-header $(pwd)/docs/html/custom-header.html --document-private-items"
