+++ src/adaptive/mod.rs
